CSS overhaul + directory moves
