+++ components/items/Code.module.css
@@ -1,0 +1,26 @@
+.code {
+    font-family: monospace;
+    font-size: 1.2em;
+    background-color: #ec3750;
+    padding: 0.5em;
+    border-radius: 5px;
+    display: block;
+    overflow: auto;
+    white-space: pre;
+    word-wrap: normal;
+    word-break: normal;
+    line-height: 1.5em;
+    max-height: 20em;
+    overflow-y: auto;
+    overflow-x: auto;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+}
+
+@media screen and (max-width: 600px) {
+    .code {
+        white-space: pre-wrap;
+    }
+}
