+++ components/items/SoonWidget.module.css
@@ -4,16 +4,22 @@
 }
 
 .quoteContainer {
+    display: -webkit-box;
+    display: -ms-flexbox;
     display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
     flex-direction: column;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
     align-items: center;
     text-align: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
     justify-content: center;
     margin: 0 auto;
     padding: 1rem;
-
-    border: 2px solid white;
-    border-radius: 10px;
 }
 
 .quoteText {
