+++ components/sections/Profile.module.css
@@ -1,0 +1,135 @@
+.profile {
+    width: 100%;
+    max-width: 800px;
+    margin: 0 auto;
+}
+
+.profileImage {
+    width: 100px;
+    height: 100px;
+    border-radius: 10%;
+}
+
+.profileHeader {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-flex: 1;
+    -ms-flex-positive: 1;
+    flex-grow: 1;
+}
+
+.profileTitle {
+    margin: 20px;
+}
+
+.socialMedia {
+    margin-right: 10px;
+}
+
+.socialMediaIcon {
+    margin: 0 10px;
+}
+
+.socialMediaIcon:hover {
+    cursor: pointer;
+    -webkit-transform: scale(1.2);
+    -ms-transform: scale(1.2);
+    transform: scale(1.2);
+}
+
+.grid {
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+    -ms-flex-wrap: wrap;
+    flex-wrap: wrap;
+}
+
+.card {
+    min-width: 300px;
+    margin: 1rem;
+    padding: 1.5rem;
+    text-align: left;
+    color: inherit;
+    text-decoration: none;
+    border: 1px solid white;
+    border-radius: 10px;
+    -webkit-transition: color 0.15s ease, border-color 0.15s ease;
+    -o-transition: color 0.15s ease, border-color 0.15s ease;
+    transition: color 0.15s ease, border-color 0.15s ease;
+}
+
+.card:hover,
+.card:focus,
+.card:active {
+    color: #ec3750;
+    border-color: #ec3750;
+}
+
+.card h2 {
+    margin: 0 0 1rem 0;
+    font-size: 1.5rem;
+}
+
+.card p {
+    margin: 0;
+    font-size: 1.2rem;
+    line-height: 1.5;
+}
+
+.button {
+    cursor: pointer;
+    padding: 0.5rem 1rem;
+    margin: 1rem 0 0;
+    font-size: 1.2rem;
+    font-weight: 600;
+    border: 0;
+    border-radius: 9999px;
+    color: white;
+    background-color: #ec3750;
+}
+
+@media (max-width: 600px) {
+    .profileHeader {
+        -webkit-box-orient: vertical;
+        -webkit-box-direction: normal;
+        -ms-flex-direction: column;
+        flex-direction: column;
+        -webkit-box-align: center;
+        -ms-flex-align: center;
+        align-items: center;
+    }
+    .profileDescription {
+        text-align: center;
+    }
+}
+
+@media (prefers-color-scheme: dark) {
+    .socialMedia img {
+        -webkit-filter: invert(1);
+        filter: invert(1);
+    }
+}
+
+@media (max-width: 600px) {
+    .grid {
+        width: 100%;
+        -webkit-box-orient: vertical;
+        -webkit-box-direction: normal;
+        -ms-flex-direction: column;
+        flex-direction: column;
+    }
+}
+
+@media (prefers-color-scheme: dark) {
+    .card {
+        border-color: #222;
+    }
+}
