+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@next/font": "13.1.4",
     "@vercel/analytics": "^0.1.8",
     "eslint": "8.32.0",
     "eslint-config-next": "13.1.4",
