--- pages/components/items/Code.module.css
@@ -1,15 +1,0 @@
-.code {
-  text-align: left; /* override default center alignment */
-  color: #ec3750;
-  border-radius: 5px;
-  padding: 0.75rem;
-  font-size: 1.1rem;
-  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-    Bitstream Vera Sans Mono, Courier New, monospace;
-  border: 1px solid #ec3750;
-}
-
-.description {
-  padding: 0.75rem;
-  font-size: 1.1rem;
-}
