--- pages/components/sections/Profile.module.css
@@ -1,111 +1,0 @@
-.profile {
-  width: 100%;
-  max-width: 800px;
-  margin: 0 auto;
-}
-
-.profileImage {
-  width: 100px;
-  height: 100px;
-  border-radius: 10%;
-}
-
-.profileHeader {
-  display: flex;
-  align-items: center;
-  flex-grow: 1;
-}
-
-.profileTitle {
-  margin: 20px;
-}
-
-.socialMedia {
-  margin-right: 10px;
-}
-
-.socialMediaIcon {
-  margin: 0 10px;
-}
-
-.socialMediaIcon:hover {
-  cursor: pointer;
-  transform: scale(1.2);
-}
-
-.grid {
-  align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
-}
-
-.card {
-  min-width: 300px;
-  margin: 1rem;
-  padding: 1.5rem;
-  text-align: left;
-  color: inherit;
-  text-decoration: none;
-  border: 1px solid white;
-  border-radius: 10px;
-  transition: color 0.15s ease, border-color 0.15s ease;
-}
-
-.card:hover,
-.card:focus,
-.card:active {
-  color: #ec3750;
-  border-color: #ec3750;
-}
-
-.card h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
-}
-
-.card p {
-  margin: 0;
-  font-size: 1.2rem;
-  line-height: 1.5;
-}
-
-.button {
-  cursor: pointer;
-  padding: 0.5rem 1rem;
-  margin: 1rem 0 0;
-  font-size: 1.2rem;
-  font-weight: 600;
-  border: 0;
-  border-radius: 9999px;
-  color: white;
-  background-color: #ec3750;
-}
-
-@media (max-width: 600px) {
-  .profileHeader {
-    flex-direction: column;
-    align-items: center;
-  }
-  .profileDescription {
-    text-align: center;
-  }
-}
-
-@media (prefers-color-scheme: dark) {
-  .socialMedia img {
-    filter: invert(1);
-  }
-}
-
-@media (max-width: 600px) {
-  .grid {
-    width: 100%;
-    flex-direction: column;
-  }
-}
-
-@media (prefers-color-scheme: dark) {
-  .card {
-    border-color: #222;
-  }
-}
