+++ styles/globals.css
@@ -1,50 +1,56 @@
 html,
 body {
-  padding: 0;
-  margin: 0;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-    Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
+    padding: 0;
+    margin: 0;
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI",
+        Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue",
+        sans-serif;
 }
 
 h1 {
-  font-size: 2rem;
-  font-weight: 600;
-  line-height: 1.15;
-  margin: 0;
+    font-size: 2rem;
+    font-weight: 600;
+    line-height: 1.15;
+    margin: 0;
 }
 
 h2 {
-  font-size: 1.5rem;
-  font-weight: 600;
-  line-height: 1.15;
-  margin: 0;
+    font-size: 1.5rem;
+    font-weight: 600;
+    line-height: 1.15;
+    margin: 0;
 }
 
 p {
-  line-height: 1.5;
-  font-size: 1.25rem;
+    line-height: 1.5;
+    font-size: 1.25rem;
 }
 
 a {
-  color: #ec3750;
-  text-decoration: none;
+    color: #ec3750;
+    text-decoration: none;
 }
 
 a:hover:after {
-  transform: scaleX(1);
-  transform-origin: bottom left;
+    -webkit-transform: scaleX(1);
+    -ms-transform: scaleX(1);
+    transform: scaleX(1);
+    -webkit-transform-origin: bottom left;
+    -ms-transform-origin: bottom left;
+    transform-origin: bottom left;
 }
 
 * {
-  box-sizing: border-box;
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
 }
 
 @media (prefers-color-scheme: dark) {
-  html {
-    color-scheme: dark;
-  }
-  body {
-    color: white;
-    background: black;
-  }
+    html {
+        color-scheme: dark;
+    }
+    body {
+        color: white;
+        background: black;
+    }
 }
